refactor(schema): split migration domain modules
