+++ packages/compiler/src/schema/index.ts
@@ -1,3 +1,23 @@
+export {
+	createCommittedMigration,
+	verifyCommittedMigration,
+	verifyCommittedMigrationChain,
+} from "./committed-migration";
+export type {
+	CommittedMigration,
+	CommittedMigrationFilesV1,
+	MigrationPlanV1,
+	MigrationStepKindV1,
+	MigrationStepV1,
+	RenameIdentityV1,
+	SchemaProjectionV1,
+} from "./contracts";
+export { createMigrationPlan } from "./migration-plan";
+export type {
+	MigrationPlanningResult,
+	NoChangesMigration,
+	PlannedMigration,
+} from "./migration-planning";
 export { loadCommittedMigration } from "./migration-files";
 export {
 	explainCommittedMigration,
