+++ quality/code-architecture.json
@@ -5,7 +5,6 @@
 	"maximumLines": 800,
 	"legacy": {
 		"packages/compiler/src/model.ts": 961,
-		"packages/compiler/src/schema-postgres.ts": 1570,
-		"packages/compiler/src/schema.ts": 1397
+		"packages/compiler/src/schema-postgres.ts": 1592
 	}
 }
